refactor!: rename app module to core
BREAKING CHANGE: Application has been renamed from 'app' to 'core' to better reflect its role as the core module of the AutoVisionAI system. This change affects all references to the application name throughout the codebase.
