--- src/autovisionai/app/loggers/__init__.py
@@ -1,3 +1,0 @@
-from autovisionai.app.loggers.app_logger import AppLogger, setup_app_logger
-
-__all__ = [AppLogger, setup_app_logger]
