+++ src/autovisionai/core/configs/local/data.yaml
