+++ src/autovisionai/core/configs/local/models.yaml
