+++ src/autovisionai/core/configs/prod/models.yaml
