+++ src/autovisionai/core/models/fast_scnn/fast_scnn_model.py
