+++ src/autovisionai/core/processing/dataset.py
@@ -6,8 +6,8 @@ import torch
 from PIL import Image
 from torch.utils.data import Dataset
 
-from autovisionai.app.configs import CONFIG
-from autovisionai.app.utils.utils import get_valid_files
+from autovisionai.core.configs import CONFIG
+from autovisionai.core.utils.utils import get_valid_files
 
 
 class CarsDataset(Dataset):
