+++ src/autovisionai/core/utils/common.py
